renamed adx to dec
