+++ cmd/azqr/scan.go
@@ -11,7 +11,6 @@ import (
 	"time"
 
 	"github.com/Azure/azqr/internal/scanners"
-	"github.com/Azure/azqr/internal/scanners/adx"
 	"github.com/Azure/azqr/internal/scanners/afd"
 	"github.com/Azure/azqr/internal/scanners/afw"
 	"github.com/Azure/azqr/internal/scanners/agw"
@@ -24,6 +23,7 @@ import (
 	"github.com/Azure/azqr/internal/scanners/cog"
 	"github.com/Azure/azqr/internal/scanners/cosmos"
 	"github.com/Azure/azqr/internal/scanners/cr"
+	"github.com/Azure/azqr/internal/scanners/dec"
 	"github.com/Azure/azqr/internal/scanners/evgd"
 	"github.com/Azure/azqr/internal/scanners/evh"
 	"github.com/Azure/azqr/internal/scanners/kv"
@@ -72,7 +72,6 @@ var scanCmd = &cobra.Command{
 	Args:  cobra.NoArgs,
 	Run: func(cmd *cobra.Command, args []string) {
 		serviceScanners := []scanners.IAzureScanner{
-			&adx.DataExplorerScanner{},
 			&afd.FrontDoorScanner{},
 			&afw.FirewallScanner{},
 			&agw.ApplicationGatewayScanner{},
@@ -85,6 +84,7 @@ var scanCmd = &cobra.Command{
 			&cog.CognitiveScanner{},
 			&cosmos.CosmosDBScanner{},
 			&cr.ContainerRegistryScanner{},
+			&dec.DataExplorerScanner{},
 			&evgd.EventGridScanner{},
 			&evh.EventHubScanner{},
 			&kv.KeyVaultScanner{},
